fix: handle optional sheets
If a sheet is marked as optional we do not expect the library to throw
an error when the sheet is not found. We expect an empty array.

From production logs it appears the Google API rate limits calls for a
spreadsheet when there has been multiple failed fetches to the same
spreadsheet. When doing this it also slows down the responses.
This is kind of failure is much more likely for optional sheet because
we're fetching a sheet that may not be there. So we're happy for a little
bit of grey failure in this case.